requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community # <--- Make sure this line is present
openai
chromadb
unstructured[md,html] # <--- And this line with extras
lxml
markdown
tiktoken
networkx
pyvis